move _session to session
