+++ cognite/neat/__init__.py
@@ -1,5 +1,5 @@
-from ._session import NeatSession
 from ._utils.auth import get_cognite_client
 from ._version import __version__
+from .session import NeatSession
 
 __all__ = ["NeatSession", "__version__", "get_cognite_client"]
