--- cognite/neat/_session/__init__.py
@@ -1,3 +1,0 @@
-from ._base import NeatSession
-
-__all__ = ["NeatSession"]
