+++ cognite/neat/session/__init__.py
@@ -1,0 +1,3 @@
+from ._base import NeatSession
+
+__all__ = ["NeatSession"]
