+++ cognite/neat/session/_base.py
