+++ cognite/neat/session/_collector.py
