+++ cognite/neat/session/_drop.py
