+++ cognite/neat/session/_explore.py
