+++ cognite/neat/session/_inspect.py
