+++ cognite/neat/session/_show.py
@@ -8,9 +8,9 @@ from pyvis.network import Network as PyVisNetwork  # type: ignore
 
 from cognite.neat._constants import IN_NOTEBOOK, IN_PYODIDE
 from cognite.neat._rules.analysis._base import RulesAnalysis
-from cognite.neat._session.exceptions import NeatSessionError
 from cognite.neat._utils.io_ import to_directory_compatible
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri, uri_display_name
+from cognite.neat.session.exceptions import NeatSessionError
 
 from ._state import SessionState
 from .exceptions import session_class_wrapper
