+++ cognite/neat/session/_template.py
