+++ cognite/neat/session/exceptions.py
